requirements.txt
asgiref==3.2.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==2.0
django-braces==1.13.0
django-crispy-forms==1.8.1
django-filebrowser==3.12.1
django-grappelli==2.14.2
django-model-utils==3.2.0
filelock==3.0.12
httplib2==0.14.0
idna==2.10
importlib-metadata==0.23
jose==1.0.0
more-itertools==7.2.0
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==19.2
Pillow==6.2.1
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyJWT==1.7.1
pyparsing==2.4.5
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
social-auth-core==3.3.3
sqlparse==0.3.0
toml==0.10.0
tox==3.14.1
Unipath==1.1
urllib3==1.25.11
virtualenv==16.7.8
zipp==0.6.0
